==> Originals/Precios/3. Proveedores más baratos.ipynb <==
import pandas as pd
import numpy as np
import re as re

Compara los precios entre proveedores y deja los tres más baratos -sin tener en cuenta los 0-. Si no hay más de un proveedor, deja “-”.

Creamos el dataframe.

Proveedores = 'J:/My Drive/Forraje/Proveedores.xlsx'

# Crear el DataFrame utilizando esas columnas
df_Proveedores = pd.read_excel(Proveedores, sheet_name='Centro')


Ponemos las listas de columnas.

lista_proveedores = ['Benavídez', 'Gallo', 'Anea', 'Pilar', 'Tortuguitas', 'Del Viso',
                     'Chica Glade', 'San Martín', 'Hipocampus', 'Mundo Mascotas', 'Yaguar',
                     'King Clor', 'Diserquim', 'Moreno', 'Roberto', 'Viejo Gets', 'Viejo Cuero', 'Mis Nietos', 
                     'Profesional Vet', 'Premier', 'Dacam', 'Proveedor', 'Saphirus']

Borramos los NaN de las columnas de proveedores y los reemplazamos con cero.

# Reemplazar NaN con 0
df_Proveedores[lista_proveedores] = df_Proveedores[lista_proveedores].fillna(0)

Convertimos a float los valores strings que puedan llegar a ser problemáticos. También borramos en el caso de que haya palabras.

# Convertir las columnas especificadas a float64
for i in lista_proveedores:
    df_Proveedores[i] = df_Proveedores[i].astype(float)


Calcula el mínimo entre todas las columnas, e ignora los cero. Si todas son cero, queda cero.

# Encuentra el mínimo excluyendo los valores de cero
primer_minimo = df_Proveedores[lista_proveedores].replace(0, np.nan).min(axis=1, skipna=True)

# Encuentra el segundo mínimo excluyendo los valores de cero
segundo_minimo = df_Proveedores[lista_proveedores].apply(lambda row: sorted(filter(lambda x: x != 0, row))[:2][1] if len([x for x in row if x != 0]) >= 2 else 0, axis=1)

# Encuentra el segundo mínimo excluyendo los valores de cero
tercer_minimo = df_Proveedores[lista_proveedores].apply(lambda row: sorted(filter(lambda x: x != 0, row))[:3][2] if len([x for x in row if x != 0]) >= 3 else 0, axis=1)


# Actualiza la columna 'Precios 1, 2 y 3' con los mínimos calculados
df_Proveedores['Precio 1'] = primer_minimo
df_Proveedores['Precio 2'] = segundo_minimo
df_Proveedores['Precio 3'] = tercer_minimo

# Columnas de diferencias.
df_Proveedores['Diferencia (2°)'] = df_Proveedores['Precio 2'] - df_Proveedores['Precio 1']
df_Proveedores['Diferencia (3°)'] = df_Proveedores['Precio 3'] - df_Proveedores['Precio 1']


# Reemplazar NaN con 0
df_Proveedores['Precio 1'] = df_Proveedores['Precio 1'].fillna(0)
df_Proveedores['Precio 2'] = df_Proveedores['Precio 2'].fillna(0)
df_Proveedores['Precio 3'] = df_Proveedores['Precio 3'].fillna(0)
df_Proveedores['Diferencia (2°)'] = df_Proveedores['Diferencia (2°)'].fillna(0)
df_Proveedores['Diferencia (3°)'] = df_Proveedores['Diferencia (3°)'].fillna(0)

# Reemplazar valores negativos por 0.
df_Proveedores.loc[df_Proveedores['Diferencia (2°)'] < 0, 'Diferencia (2°)'] = '-'
df_Proveedores.loc[df_Proveedores['Diferencia (3°)'] < 0, 'Diferencia (3°)'] = '-'

Ahora lo que hace es llenar las columnas "Proveedor 1, 2 y 3" con el proveedor que tiene más barato el producto.

# Iteramos sobre cada fila del DataFrame
for indice, fila in df_Proveedores.iterrows():
    # Iteramos sobre cada columna de lista_proveedores
    for columna in lista_proveedores:
        # Comparamos el valor de "Precio" con el valor en la columna actual
        if fila["Precio 1"] == 0:
            df_Proveedores.at[indice, "Proveedor 1"] = "-"
        elif fila["Precio 1"] == fila[columna]:
            # Si coinciden, actualizamos la columna "Proveedores" con el nombre de la columna actual
            df_Proveedores.at[indice, "Proveedor 1"] = columna
            # Salimos del bucle interno
            break


# Iteramos sobre cada fila del DataFrame
for indice, fila in df_Proveedores.iterrows():
    # Iteramos sobre cada columna de lista_proveedores
    for columna in lista_proveedores:
        # Comparamos el valor de "Precio" con el valor en la columna actual
        if fila["Precio 2"] == 0:
            df_Proveedores.at[indice, "Proveedor 2"] = "-"
        elif fila["Precio 2"] == fila[columna]:
            # Si coinciden, actualizamos la columna "Proveedores" con el nombre de la columna actual
            df_Proveedores.at[indice, "Proveedor 2"] = columna
            # Salimos del bucle interno
            break

# Iteramos sobre cada fila del DataFrame
for indice, fila in df_Proveedores.iterrows():
    # Iteramos sobre cada columna de lista_proveedores
    for columna in lista_proveedores:
        # Comparamos el valor de "Precio" con el valor en la columna actual
        if fila["Precio 3"] == 0:
            df_Proveedores.at[indice, "Proveedor 3"] = "-"
        elif fila["Precio 3"] == fila[columna]:
            # Si coinciden, actualizamos la columna "Proveedores" con el nombre de la columna actual
            df_Proveedores.at[indice, "Proveedor 3"] = columna
            # Salimos del bucle interno
            break

df_Proveedores['Precio 1'] = df_Proveedores['Precio 1'].replace(0, '-')
df_Proveedores['Precio 2'] = df_Proveedores['Precio 2'].replace(0, '-')
df_Proveedores['Precio 3'] = df_Proveedores['Precio 3'].replace(0, '-')
df_Proveedores['Diferencia (2°)'] = df_Proveedores['Diferencia (2°)'].replace(0, '-')
df_Proveedores['Diferencia (3°)'] = df_Proveedores['Diferencia (3°)'].replace(0, '-')

# Se importa Database.

Database = 'J:/My Drive/Forraje/Marcas y subcategorías.xlsx'

# Crear el DataFrame utilizando esas columnas
df_Marcas = pd.read_excel(Database)

# Ordenar según la columna 'Descripcion' en orden ascendente.

df_Marcas = df_Marcas.sort_values(by='Descripcion')

df_Marcas

# Resetear index del df.
df_Marcas = df_Marcas.reset_index(drop=True)

# Importar datos de marcas y subcategorias.

print(len(df_Marcas))
print(len(df_Proveedores))

if len(df_Marcas) == len(df_Proveedores):
    df_Proveedores['Marca'] = df_Marcas['Marca']
    df_Proveedores['Subcategoria'] = df_Marcas['Subcategoria']

# Nombres de las columnas en orden.
Orden = ['Codigo', 'Descripcion', 'Categoria', 'Subcategoria', 'Marca',
         'Proveedor 1', 'Precio 1', 
         'Proveedor 2', 'Precio 2', 'Diferencia (2°)',
         'Proveedor 3', 'Precio 3', 'Diferencia (3°)']

# Ordena el DataFrame
df_Proveedores = df_Proveedores[Orden]

df_Proveedores

Creamos un .xlsx

# Guardar el DataFrame en un archivo.
Ruta_1 = f"J:/My Drive/Forraje/Baratos.xlsx"
df_Proveedores.to_excel(Ruta_1, index=False)
